Theta tweaks on 0.3.5.1     - JobSource does not filter by queued_launch; any job is eligible to run in a launcher job, regardless of which it was reserved for.     - dag.add_dependency does not check for circular dependencies; this is more dangerous but a lot faster (when DAG gets big, the python code is too slow)     - launcher.py: can set 0 threads; then Transitions DO NOT run alongside launcher.  This is important for transitions that require outbound internet access     - mpi_ensemble: serial job mode does not quit prematurely if there are still processable jobs along the way.     - job packer counts AWAIITNG_PARENTS jobs as eligible to be packed for queue launch     - dummy QueuedLaunch is now 256 nodes and 6 hours     - The Balsam service creates 5 transition threads on startup, by default, instead of 1

+++ balsam/launcher/launcher.py
@@ -418,13 +418,16 @@ def main(args):
     Launcher = MPILauncher if job_mode == 'mpi' else SerialLauncher
     
     try:
-        transition_pool = transitions.TransitionProcessPool(nthread, wf_filter)
+        if nthread > 0:
+            transition_pool = transitions.TransitionProcessPool(nthread, wf_filter)
+        else:
+            transition_pool = None
         launcher = Launcher(wf_filter, timelimit_min, gpus_per_node)
         launcher.run()
     except:
         raise
     finally:
-        transition_pool.terminate()
+        if transition_pool is not None: transition_pool.terminate()
         logger.info("Exit: Launcher exit graceful\n\n")
 
 def get_args(inputcmd=None):
